Support streaming data to ffmpeg/ffprobe's stdin
It is common for intents in Android to provide data as a uri,
which may not be directly resolvable to a file path. Instead,
a uri can be resolved as an input stream. To allow ffmpeg and
ffprobe to consume such data, this change adds support for streaming
data to the forked process' stdin.